refactor(app): extract share picker lifecycle
