+++ src/app.rs
@@ -49,6 +49,7 @@ pub mod runtime_loop;
 pub mod runtime_media_events;
 pub mod runtime_startup;
 pub mod share_picker;
+pub mod share_picker_input;
 pub mod status_actions;
 pub mod status_input;
 pub mod status_projection;
